requirements.txt
faiss_cpu
jieba
numpy
openai
py2neo
PyYAML
sentence_transformers
tqdm
